pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "safeanchor"
dynamic = ["version"]
description = "Preventing cumulative safety erosion in continual domain adaptation of large language models"
readme = "README.md"
requires-python = ">=3.10,<3.13"
authors = [
    {name = "Anonymous Author(s)", email = "anonymous@example.com"},
]
maintainers = [
    {name = "Anonymous Author(s)", email = "anonymous@example.com"},
]
keywords = [
    "AI safety",
    "continual learning",
    "large language models",
    "safety alignment",
    "LoRA",
    "domain adaptation",
    "Fisher information",
    "orthogonal gradient projection",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

dependencies = [
    "torch>=2.1.0",
    "numpy>=1.24.0,<2.0.0",
    "transformers>=4.38.0",
    "datasets>=2.18.0",
    "tokenizers>=0.15.0",
    "peft>=0.9.0",
    "accelerate>=0.28.0",
    "safetensors>=0.4.2",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "wandb>=0.16.0",
    "tensorboard>=2.15.0",
    "tqdm>=4.66.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "huggingface-hub>=0.21.0",
    "einops>=0.7.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "scipy>=1.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "pytest-timeout>=2.2.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "ipython>=8.19.0",
    "ipdb>=0.13.13",
]
docs = [
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]
all = [
    "safeanchor[dev,docs]",
]

[project.urls]
Homepage = "https://anonymous.4open.science/r/SafeAnchor"
Repository = "https://anonymous.4open.science/r/SafeAnchor"
Issues = "https://anonymous.4open.science/r/SafeAnchor"

[project.scripts]
safeanchor-train = "safeanchor.training.trainer:main"
safeanchor-eval  = "safeanchor.evaluation.evaluator:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["safeanchor*"]

[tool.setuptools.dynamic]
version = {attr = "safeanchor.__version__.__version__"}

[tool.setuptools.package-data]
"safeanchor" = ["py.typed", "*.yaml", "*.json"]

# ─────────────────────────────────────────────────────────────────────────────
# RUFF
# ─────────────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]
exclude = [
    ".git", ".mypy_cache", ".ruff_cache", ".venv", "venv",
    "__pycache__", "build", "dist",
]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "C4", "SIM", "PTH"]
ignore = ["E501", "PLR0913", "PLR2004", "N812", "N806"]

[tool.ruff.lint.isort]
known-first-party = ["safeanchor"]
lines-after-imports = 2

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["PLR2004", "S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

# ─────────────────────────────────────────────────────────────────────────────
# MYPY
# ─────────────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "transformers.*", "datasets.*", "wandb.*", "hydra.*",
    "omegaconf.*", "torch.*", "numpy.*", "peft.*",
    "accelerate.*", "scipy.*", "tqdm.*",
]
ignore_missing_imports = true

# ─────────────────────────────────────────────────────────────────────────────
# PYTEST
# ─────────────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "-ra",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=xml",
    "--cov-fail-under=70",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests requiring GPU",
    "integration: marks integration tests",
    "unit: marks unit tests",
]
filterwarnings = ["ignore::DeprecationWarning", "ignore::UserWarning"]
timeout = 300

[tool.coverage.run]
source = ["src"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]
fail_under = 70
show_missing = true
